Development

+++ src/actions/favorites.js
@@ -1,37 +1,52 @@
 import firebase from 'firebase';
 import _ from 'lodash';
 import { AsyncStorage } from 'react-native';
-import { Actions } from 'react-native-router-flux';
 import { store } from '../App';
 import {
 	FAVORITE_CREATE, FAVORITE_CREATE_FAIL, FAVORITE_DELETE,
 	FAVORITE_FETCH_SUCCESS, FAVORITE_FETCH_FAIL,
 	ERROR,
-	CLR_SEARCH,
 	LINES_FETCH, LINE_ADD, LINE_REMOVE
 } from './types';
+import { generateUid } from '../components/helpers';
 
 // Stops
 
 export const favoriteCreate = ({ busStop, id }) => {
 	const { currentUser } = firebase.auth();
 	return (dispatch) => {
-		const fbRef = firebase.database().ref(`/users/${currentUser.uid}/favorites`);
-		fbRef.orderByChild('id').equalTo(id).once('value', (snapshot) => {
-			if (snapshot.val() == null) {
-				fbRef.push({ busStop, id })
-					.then(() => {
-						dispatch({ type: FAVORITE_CREATE });
-						dispatch({ type: CLR_SEARCH });
-						Actions.dashboard({ type: 'reset' });
-					}, (error) => {
-						window.log('favoriteCreate error: ', error);
-						favoriteCreateFail(dispatch);
-					});
-			} else {
-				dispatch(favoriteDelete(id));
-			}
-		});
+		if (currentUser.isAnonymous) {
+			return AsyncStorage.getItem('minahallplatser-favorites').then((dataJson) => {
+				const favorites = JSON.parse(dataJson) || {};
+				let exists = false;
+				_.forEach(favorites, (item, key) => {
+					if (item.id === id) {
+						exists = true;
+						delete favorites[key];
+					}
+				});
+				if (!exists) {
+					favorites[generateUid()] = { busStop, id };
+				}
+				dispatch({ type: FAVORITE_FETCH_SUCCESS, payload: favorites })
+				return AsyncStorage.setItem('minahallplatser-favorites', JSON.stringify(favorites));
+			});
+		} else {
+			const fbRef = firebase.database().ref(`/users/${currentUser.uid}/favorites`);
+			fbRef.orderByChild('id').equalTo(id).once('value', (snapshot) => {
+				if (snapshot.val() == null) {
+					fbRef.push({ busStop, id })
+						.then(() => {
+							dispatch({ type: FAVORITE_CREATE });
+						}, (error) => {
+							window.log('favoriteCreate error: ', error);
+							favoriteCreateFail(dispatch);
+						});
+				} else {
+					dispatch(favoriteDelete(id));
+				}
+			});
+		}
 	};
 };
 
@@ -49,19 +64,24 @@ export const favoriteGet = (currentUser) => {
 					dispatch({ type: FAVORITE_FETCH_SUCCESS, payload: favorites });
 				} else {
 					window.log('Did not find favorites locally');
+					if (currentUser.isAnonymous) {
+						return dispatch({ type: FAVORITE_FETCH_SUCCESS, payload: [] })
+					}
 				}
-				firebase.database().ref(`/users/${currentUser.uid}/favorites`)
-				.on('value', snapshot => {
-					AsyncStorage.setItem('minahallplatser-favorites', JSON.stringify(snapshot.val()));
-					dispatch({ type: FAVORITE_FETCH_SUCCESS, payload: snapshot.val() });
-				}, (error) => {
-					window.log('favoriteGet error: ', error);
-					dispatch({
-						type: FAVORITE_FETCH_FAIL,
-						payload: { loading: false }
+				if (!currentUser.isAnonymous) {
+					firebase.database().ref(`/users/${currentUser.uid}/favorites`)
+					.on('value', snapshot => {
+						AsyncStorage.setItem('minahallplatser-favorites', JSON.stringify(snapshot.val()));
+						dispatch({ type: FAVORITE_FETCH_SUCCESS, payload: snapshot.val() });
+					}, (error) => {
+						window.log('favoriteGet error: ', error);
+						dispatch({
+							type: FAVORITE_FETCH_FAIL,
+							payload: { loading: false }
+						});
+						dispatch({ type: ERROR, payload: 'Kunde inte ladda dina favoriter.' });
 					});
-					dispatch({ type: ERROR, payload: 'Kunde inte ladda dina favoriter.' });
-				});
+				}
 				AsyncStorage.getItem('minahallplatser-lines').then((localLines) => {
 					const lines = JSON.parse(localLines);
 					if (lines) {
@@ -70,14 +90,16 @@ export const favoriteGet = (currentUser) => {
 						window.log('Did not find lines locally');
 					}
 				});
-				firebase.database().ref(`/users/${currentUser.uid}/lines`)
-				.once('value', snapshot => {
-					AsyncStorage.setItem('minahallplatser-lines', JSON.stringify(snapshot.val()));
-					dispatch({ type: LINES_FETCH, payload: snapshot.val() });
-				}, (error) => {
-					window.log('lines error: ', error);
-					dispatch({ type: ERROR, payload: 'Kunde inte ladda dina sparade linjer.' });
-				});
+				if (!currentUser.isAnonymous) {
+					firebase.database().ref(`/users/${currentUser.uid}/lines`)
+					.once('value', snapshot => {
+						AsyncStorage.setItem('minahallplatser-lines', JSON.stringify(snapshot.val()));
+						dispatch({ type: LINES_FETCH, payload: snapshot.val() });
+					}, (error) => {
+						window.log('lines error: ', error);
+						dispatch({ type: ERROR, payload: 'Kunde inte ladda dina sparade linjer.' });
+					});
+				}
 			}).catch((err) => {
 				window.log(err);
 			});
@@ -128,27 +150,43 @@ export const favoriteLineToggle = ({ sname, direction }) => {
 			exists = true;
 			dispatch({ type: FAVORITE_CREATE_FAIL });
 		}
-		if (!exists) {
-			window.log('favoriteLineToggle push:', line);
-			dispatch({ type: LINE_ADD, payload: line });
-			fbRef.push(line)
-			.catch((error) => {
-				window.log('favoriteLineToggle push error:', error);
-				favoriteCreateFail(dispatch);
+		if (currentUser.isAnonymous) {
+			AsyncStorage.getItem('minahallplatser-lines').then((localLines) => {
+				const lines = JSON.parse(localLines) || {};
+				if (!exists || !lines) {
+					lines[generateUid()] = line;
+					AsyncStorage.setItem('minahallplatser-lines', JSON.stringify(lines));
+					dispatch({ type: LINE_ADD, payload: line });
+				} else {
+					const key = _.findKey(lines, (item) => item == line);
+					delete lines[key];
+					AsyncStorage.setItem('minahallplatser-lines', JSON.stringify(lines));
+					dispatch({ type: LINE_REMOVE, payload: line });
+				}
 			});
 		} else {
-			window.log('favoriteLineToggle remove:', line);
-			dispatch({ type: LINE_REMOVE, payload: line });
-			fbRef.once('value', snapshot => {
-				const fbLines = snapshot.val();
-				_.forEach(fbLines, (item, key) => {
-					if (item === line) {
-						fbKey = fbRef.child(key);
-					}
+			if (!exists) {
+				window.log('favoriteLineToggle push:', line);
+				dispatch({ type: LINE_ADD, payload: line });
+				fbRef.push(line)
+				.catch((error) => {
+					window.log('favoriteLineToggle push error:', error);
+					favoriteCreateFail(dispatch);
 				});
-				fbKey.remove()
-				.catch((error) => window.log(error));
-			});
+			} else {
+				window.log('favoriteLineToggle remove:', line);
+				dispatch({ type: LINE_REMOVE, payload: line });
+				fbRef.once('value', snapshot => {
+					const fbLines = snapshot.val();
+					_.forEach(fbLines, (item, key) => {
+						if (item === line) {
+							fbKey = fbRef.child(key);
+						}
+					});
+					fbKey.remove()
+					.catch((error) => window.log(error));
+				});
+			}
 		}
 	};
 };
